change int to double, add exceptions message

+++ hw2/hw2.3/IStack.cs
@@ -4,8 +4,8 @@ namespace hwTwoDotThree
     {
         bool IsEmpty();
 
-        void Push(int value);
+        void Push(double value);
 
-        int Pop();
+        double Pop();
     }
 }
